MoKeeSetupWizard: offer to use WiFi to sign into Google Services
Google setup wizard prompts for WiFi only if a data connection is not
available, and this is not the desired behavior. So let's prompt the
user whether they wish to connect to a WiFi network after selecting
their language. But we do this only once during the setup process.

Change-Id: I70cc0cecc9861595d1b672cd116b1fee2720efe3
Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:
	src/com/mokee/setupwizard/ui/SetupWizardActivity.java