SetupWizard: add a MK-specific setup-complete settings key
Android uses Settings.Global.DEVICE_PROVISIONED and
Settings.Secure.USER_SETUP_COMPLETE and the system responds to those
database entries, such as allowing the keyguard to be visible, or maybe
pulling the statusbar down.

Since we still depend on some of Google's setup wizard, it also keys off
the same table rows and it may decide that the device is now provisioned
when we still haven't finished our own setup wizard, so let's add
another key that we have control over to show the real state of OUR
setup wizard.

Ticket: CYNGNOS-2431

Change-Id: Ic21689625f6d80e209161762ace741918b972aba
Signed-off-by: Roman Birg <roman@cyngn.com>